[native] Stop hiding keyboard when highlightable items are pressed while keyboard is up
Summary:
I'm working on [this task](https://www.notion.so/commapp/Make-it-possible-to-interact-with-MessageList-while-keyboard-is-up-1b4c5589b5914297b35629147a1e13df), where the goal is to allow the user to interact with the `MessageList` while the keyboard is up. As part of that, I'm getting rid of this logic that hides the keyboard instead of highlighting a selected component.

This diff doesn't actually prevent the keyboard from dismissing, as other components are also triggering a dismiss when something in the `MessageList` is pressed (will be addressed in future diffs). However, this diff does make it so the highlightable items are highlighted when they are pressed while the keyboard is up.

Test Plan:
I tested this whole stack together by playing around with the `MessageList` while the keyboard was up.
- I made sure that pressing outside of a highlightable component led to the keyboard being dimissed
- I made sure that the keyboard did not get dismissed when I activated an overlay by highlighting a highlightable component
- I made sure that pressing a navigational item led to the navigation occurring, as well as the keyboard being dismissed
- I tested on both iOS and Android

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1623