[native] Wrap MessageList GestureTouchableOpacitys without TouchableWithoutFeedback
Summary:
Before D164, we were using `react-native` Touchables in both the parent `Message` component, as well as the highlightable components `InlineMultimedia` and `InnerTextMessage`.

The Touchable in `Message` is meant to detect a press that occurs in `Message` but outside the highlightable components, and to dismiss the keyboard when such a touch is detected. Before D164, if a touch occured in the highlighteable components, `react-native` would not send this touch to the Touchable in `Message`.

But D164 moved the highlighteable components to use `GestureTouchableOpacity`, which is based on `GestureHandler`s from `react-native-gesture-handler`. After this change, a touch on the highlighteable components would still propagate to the Touchable in `Message`.

As part of [this task](https://www.notion.so/commapp/Make-it-possible-to-interact-with-MessageList-while-keyboard-is-up-1b4c5589b5914297b35629147a1e13df), where I am looking to make it possible to interact with the `MessageList` while the keyboard is up, we need to prevent touches on highlighteable components from propagating to the Touchable in `Message`, since that Touchable will dismiss the keyboard.

I explored several approaches to this work, and ultimately settled on the approach in this diff, which is to wrap the `GestureTouchableOpacity`s with a unused `TouchableWithoutFeedback` (no `onPress`). The inner `TouchableWithoutFeedback` catch the press and prevent it from propagating to the Touchable in `Message`.

Test Plan: I made sure that pressing a highlightable component would not trigger the Touchable in `Message`

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1624