[native] Simplify pressability rules in OverlayNavigator
Summary:
First, some history. Initially pressability was always enabled, but then in 2f70ec27957cd1157644cbf2ac7ec9fd6efc75f0 I made it so all inactive routes had `pointersEvents="none"`. The reason for this was that I wanted to make it so routes that are dismissing (which are rendered over the active route) would not capture touches, and instead the touches would propagate to the active route below the dismissing route.

Then I wanted to make an exception for `ActionResultModal`, which just displays a message on screen temporarily, and then dismisses itself. I didn't want that screen to block touches, so I added a `preventPresses` param that could be configured on a per-screen basis.

When I implemented `preventPresses`, I did it in a very convoluted way, mostly because I didn't recall my initial intention for the `pointerEvents` logic in 2f70ec27957cd1157644cbf2ac7ec9fd6efc75f0. I tried to preserve the property that inactive, non-dismissing routes (ie. routes below the current active route) had `pointerEvents="none"`. But that property was never intentional... it was just a side effect of the work in 2f70ec27957cd1157644cbf2ac7ec9fd6efc75f0 to disable dismissing routes (which are also inactive).

Anyways, back to the present day. In my work on [this task](https://www.notion.so/commapp/Make-it-possible-to-interact-with-MessageList-while-keyboard-is-up-1b4c5589b5914297b35629147a1e13df), I realized that a side effect of setting `pointerEvents` to `"none"` on a `View` was that if the keyboard was up and the active `TextInput` was inside this `View`, the keyboard would be dismissed. This meant that whenever a new overlay was displayed, it would dismiss the keyboard.

I concluded that inactive, non-dismissing `View`s should not get `pointerEvents="none"`. In terms of pressability, it actually makes no difference, since these `View`s are always rendered below another `View` with the same bounds. But it avoids the keyboard dismissal issue.

Under this new model, the logic for `preventPresses` could be simplified. We no longer have to make sure that pressability is not disabled for the route below the `preventPresses` route, since pressability is always enabled for that route.

Test Plan: I tested various overlays, and made sure that displaying a new overlay no longer dismissed the keyboard

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1627