[native] Keep messageListVerticalBounds updated while keyboard is up
Summary:
As part of [this task](https://www.notion.so/commapp/Make-it-possible-to-interact-with-MessageList-while-keyboard-is-up-1b4c5589b5914297b35629147a1e13df), I made it so the user can open overlays while the keyboard is up. But because we weren't updating `messageListVerticalBounds` while the keyboard was up, tooltips were getting rendered in the wrong direction.

This diff makes it so we keep `messageListVerticalBounds` updated, even while the keyboard is up.

Test Plan: Make sure that tooltips open in the correct directions while the keyboard is up. Test with message scrolled near top and message scrolled near bottom

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1629