[server] Add indexes to updates table to handle update replacement logic
Summary: This is to improve the performance of the update replacement logic. This particular diff addresses part of [this larger Notion issue](https://www.notion.so/commapp/Socket-crash-loop-from-creating-sidebar-in-OG-testers-thread-aac2274205394e748d0148cf049b975e), but it might be better to read this [more specific Notion issue](https://www.notion.so/commapp/Update-replacement-is-expensive-for-new-sidebar-in-large-thread-4fd267f71b754cc18562e3e622a0d52a) that covers a separate problem, since it describes the update replacement logic in a bit more depth.

Test Plan: Make sure they are created in PHPMyAdmin. Run queries individually

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1678