[server] Don't recreate relationship rows on child thread creation
Summary: We can assume that all the members of the parent already have relationship rows with one another. This assumption used to be considered before D1021. More details in this [Notion task](https://www.notion.so/commapp/Update-replacement-is-expensive-for-new-sidebar-in-large-thread-4fd267f71b754cc18562e3e622a0d52a). This diff is the primary solution for the larger issue described in [this other Notion issue](https://www.notion.so/commapp/Socket-crash-loop-from-creating-sidebar-in-OG-testers-thread-aac2274205394e748d0148cf049b975e).

Test Plan: In my local environment I have an OG SquadCal testers thread with >100 members. I tested creating a sidebar in this thread and made sure it didn't time out anymore

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1679