[server] Avoiding dispatching UPDATE_USER if relationships already exist
Summary: In `commitMembershipChangeset` we create `UPDATE_USER` updates for any relationships we set in the DB, even if we didn't change the row. This is wasteful since those updates get propagated to individual clients. We should check if the row needs to be changed before creating these updates.

Test Plan: I tested this diff without D1679 and found this decreased the time it took for the relationship logic to be handled from several minutes to just 20s. It still took 20s to run the huge `SELECT` query in this diff, but that was way faster than doing the corresponding `DELETE FROM updates` query in `createUpdates` (for the update replacement logic)

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1680