[server] Always return JOIN_THREAD on thread creation retry
Summary:
This diff is one result of my investigation into [this Notion issue](https://www.notion.so/commapp/Socket-crash-loop-from-creating-sidebar-in-OG-testers-thread-aac2274205394e748d0148cf049b975e).

The issue is that the client relies on a `JOIN_THREAD` update to update the `ThreadStore`, but will proceed with message creation as soon as the server returns success here, and will end up causing weird behavior because there's a message in the `MessageStore` with a realized `threadID` that is missing from the `ThreadStore`. Ideally the client would be more careful before sending the message – see [this Notion task](https://www.notion.so/commapp/Message-creation-can-proceed-even-when-new-thread-failed-to-merge-into-ThreadStore-01fa604573104f0a88abbf22ca6ad1ce).

Right now the server is not sending down a `JOIN_THREAD` on retry because of my mistaken assumption in [this Notion task](https://www.notion.so/commapp/Handle-duplicate-sidebar-creation-by-calling-joinThread-177056d1f17b4932bd3418b8e49e338e) that the `JOIN_THREAD` would be sent through the socket mechanism. See [this Notion task](https://www.notion.so/commapp/Deliver-return-updates-via-socket-when-request-fails-9ede3fe7107847069e80fdba53775646) for why the update doesn't come through the socket mechanism.

This diff sort of fixes the issue by making sure the `JOIN_THREAD` is delivered on retry. That said, if the state check mechanism delivers the `RawThreadInfo` the messages will end up missing and the UI for retrying will disappear. ([This Notion issue](https://www.notion.so/commapp/State-check-mechanism-should-include-messages-for-missing-threads-f2e681857e324c6f83afba72ca2d7176) tracks fixing that.)

Test Plan: I tested this diff without D1680 or D1681, and with the state check mechanism turned off. After trying to create a sidebar, and then waiting five minutes for the `createThread` to resolve, I was able to successfully retry the thread creation. (If I didn't wait five minutes, the messages would not be returned, since `createMessages` is called after `commitMembershipChangeset` in `createThread`)

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1681