[native] Set transition state when opening sidebar and unset when transition finishes
Summary: When a transition starts, we set the state to hide a message. When a screen is focused we set the state to null so that the message is displayed.

Test Plan: Open message tooltip and navigate to a sidebar. When parent thread is moved to the left it should not render the source message.

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1688