[server] Fix relationship creation for new threads
Summary: I accidentally broke this in D1680. The code currently only updates the relationship if it already exists but has a lower `status` than requested. This diff makes sure that the relationship is always created when it doesn't yet exist.

Test Plan: Manually deployed this fix to prod and verified it solved the issue

Reviewers: atul

Reviewed By: atul

Subscribers: KatPo, palys-swm, Adrian

Differential Revision: https://phabricator.ashoat.com/D1695