[native] Determine message target position
Summary: The target position will be used by setting `bottom` style property of a View that will contain `InnerMessage`s.

Test Plan: Modify text message tooltip button to call this hook, introduce a new View that contains `InnerTextMessage` and set bottom style of the `View` to the value returned. The result should be that text message is displayed at the same position that it will have in the sidebar. Open the tooltip, navigate to sidebar and verify that the positions are the same (for both pending and existing sidebar).

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1696