[native] Get rid of drawable-v21 res folder on Android
Summary: React Native 0.64 no longer supports Android 4. The new `minSdkVersion` is 21, so we can move `drawable-v21/splash_header.xml` to `drawable` so that it's used in all cases, and delete the old `drawable/splash_header.xml` that was only being used for SDK versions 20 and lower.

Test Plan: compile and run Android and make sure the splash header looks right

Reviewers: atul, palys-swm

Reviewed By: atul

Subscribers: KatPo, Adrian

Differential Revision: https://phabricator.ashoat.com/D1747