[native] Render message list when component mounts even if the overlay is visible
Summary: Currently, when the overlay is visible, we freeze the message list. We would like to animate the source message when we open a sidebar from a tooltip and for that to look nice, we need not to show a loading indicator and we need to render all the messages but the source (which will be animated). This diff makes it so that when the component mounts, we allow setting list data even when an overlay is present: the data might be already computed because we copy message heights when we create a new height measurer.

Test Plan: Tested with other diffs. At this point there should be no visible change because tooltips do not measure message list.

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1757