[native] Re-add linkPressActive property
Summary: Before the previous diff, there was one property named `linkPressActive` that blocked link presses and was active while the link `Alert` is up. In the last diff I renamed it to `linkModalActive`, and in this diff I am re-introducing `linkPressActive`. But this time `linkPressActive` is tracking whether the press gesture is active rather than tracking whether the link `Alert` is up.

Test Plan: Tested in combination with following diffs

Reviewers: palys-swm, atul

Reviewed By: palys-swm

Subscribers: KatPo, Adrian

Differential Revision: https://phabricator.ashoat.com/D1776