[native] Disable TextMessage swipeability while link modal is active
Summary: I noticed it was possible on Android to continue the swipe gesture after the `Alert` for a link press was activated. This diff fixes that issue by disabling swipeability while the `Alert` is active.

Test Plan: Make sure I can't scroll after the `Alert` for a link press is shown

Reviewers: palys-swm, atul

Reviewed By: palys-swm

Subscribers: KatPo, Adrian

Differential Revision: https://phabricator.ashoat.com/D1779