Add username and isViewer to ThreadInfo.members
Summary:
This diff changes `ThreadInfo.members` from an array of `MemberInfo` to an array of `RelativeMemberInfo`. `RelativeMemberInfo` has the same properties as `MemberInfo`, but also includes `username` and `isViewer`, which are determined from `viewerID` and `userInfos` from the Redux store.

Note that `ThreadInfo` is a client-only abstraction, and as such changing it does not have the same complexity as changing `RawThreadInfo`. (No need to worry about state check, and the server doesn't have to check `codeVersions` when delivering threads to the client.)

Test Plan: Flow, also tested in combination with the following diff

Reviewers: palys-swm, atul

Reviewed By: atul

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1786