[lib] Include other user in all sidebars inside PERSONAL threads
Summary:
`PERSONAL` threads only have two users, so it makes sense to always include the other user in the sidebar.

I considered making this change on the server side since we could deploy it faster, but then I considered that we want to support sidebars for both thin and thick threads. Thick threads won't be able to leverage the server to add the other user at creation time, so it makes sense to handle this on the client.

Test Plan: Create a sidebar in a `PERSONAL` thread from a message that the viewer authored. Make sure the other user is added to the sidebar at creation time

Reviewers: palys-swm, atul

Reviewed By: atul

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1787