[native] Allow thread search cancel button to work in activating state
Summary:
By calling `onSearchBlur` from `onSearchCancel`, we prevent the cancel button from working in the `activating` state. This makes sense for blur (in case a blur and refocus occurs during `activating`), but doesn't make sense for the cancel button.

This diff makes the cancel button work in the `activating` state.

Test Plan: Test cancel and blur of thread search separately on Android emulator

Reviewers: palys-swm, atul

Reviewed By: palys-swm

Subscribers: Adrian, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D1796