[native] Allow use of the Android back button to dismiss thread search
Summary:
We only want this to work when `ChatThreadList` is in focus.

Also worth noting that two back button presses are necessary: the first to dismiss the keyboard, and the second to dismiss the search interface. As far as I understand, this is unavoidable on Android, since the keyboard always responds to the back button if it's up.

Test Plan:
Tested on Android emulator:
1. Make sure that we can dismiss thread search with the Android back button
2. Make sure that behavior is unchanged if the `ChatThreadList` is out of focus
3. Make sure the app still correctly backgrounds if the Android back button is pressed while the `ChatThreadList` is focused but the thread search is not active

Reviewers: palys-swm, atul

Reviewed By: palys-swm

Subscribers: Adrian, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D1797