[native] Redo SwipeableMessage animation to match designs
Summary: This diff reworks the `SwipeableMessage` animation to match the [new designs](https://www.figma.com/file/L675ETKDnGaSwlpZAw4MIC/Mobile-UX-%2F-UI-Design?node-id=2%3A8). It doesn't yet add the new swipe-to-create-sidebar functionality, and the color is work (will be addressing in a later diff).

Test Plan: Play around with the new animation, including with a release build on a phone

Reviewers: palys-swm, atul

Reviewed By: palys-swm

Subscribers: Adrian, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D1800