[native] Introduce universal TextInput component
Summary: We'll avoid using the core `TextInput` from `react-native` in favor of using this new component. This diff is just a simple refactor, but the following refactor will make sure we set `keyboardAppearance` based on the active theme.

Test Plan: Play around with existing `TextInput`s and make sure behavior has not changed

Reviewers: palys-swm, atul

Reviewed By: atul

Subscribers: Adrian, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D1814