[native] Use expo-haptics for haptic feedback
Summary:
Right now we use a mix of `react-native-in-app-message`'s `TapticFeedback` for iOS and the `react-native` `Vibration` API for Android.

This diff moves us to using `expo-haptics`, which seems like a better bet in terms of being well-maintained, and allows us to use more granular APIs. In particular, we're able to use `Haptics.ImpactFeedbackStyle.Heavy` for the sidebar threshold in `SwipeableMessage`.

Test Plan: Test `SwipeableMessage` on both iOS and Android physical devices

Reviewers: atul, palys-swm

Reviewed By: atul

Subscribers: Adrian, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D1817