[native] Animate tooltip header when navigating to a sidebar
Summary: Include tooltip header in animated view and animate the opacity depending on the reason of tooltip close

Test Plan: Test robotext, text and multimedia message: open a tooltip and test closing it and navigating to sidebar. Check if in the first case, nothing is changed and in the second if the header is animated with the message.

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: Adrian, atul, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D1857