[native] Align tooltip animation duration with navigation
Summary: The navigation duration differs from platform to platform, so to make sure that the tooltip navigation looks correctly, we check what's the duration for the given platform and use the result. The value for `spring` animation (250) was estimated based on testing.

Test Plan: Tested on iOS, Android 8.1 and 9.0 simulators in release mode.

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: Adrian, atul, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D1867