[lib] urlEncode robotext generated from entry text
Summary:
I noticed a bug where creating a calendar entry with the character `%` would cause both the app and website to crash when rendering the corresponding message robotext.

The issue is that we are calling `decodeURI` on user-generated text that has not been wrapped in `encodeURI`. We wrap other user-generated text with `encodeURI`, but forgot to do the same for calendar entries.

This diff solves the issue by calling `encodeURI` in the appropriate places.

Test Plan: Make sure I can render a thread with robotext about the creation of an event with the text `%`

Reviewers: atul, palys-swm

Reviewed By: atul

Subscribers: Adrian, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D1873