[server] Dedup lists of users in createThread and updateThread
Summary:
If a client mistakenly fails to dedup the list of users it passes to the server, the server should handle that deduping itself.

This diff addresses part 2 of [this Notion task](https://www.notion.so/commapp/Specifying-userID-twice-in-initialMemberIDs-to-createThread-leads-to-crash-062791c5d498470b987b4592187d8548).

Test Plan: I modified the client to erroneously duplicate all initial members, and make sure the server deduped them by adding logging statements in `createThread`

Reviewers: atul

Reviewed By: atul

Subscribers: palys-swm, Adrian, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D1894