[lib] Fix blocking preventing messaging
Summary:
This diff addresses [this Notion issue](https://www.notion.so/commapp/Bug-Blocking-a-user-in-a-two-person-thread-55d08e2ee0174376bbf982597fe54f36).

You should be prevented from sending a message to a user who you have blocked, or who has blocked you. This does not apply to "thin threads" (threads in communities) or group threads. It only applies to one-on-one threads, such as `threadTypes.PERSONAL`.

This behavior was broken after I switched all non-community threads to be children of `Genesis`. Because children of Genesis appear to be both "thin threads" and group threads, the blocking permission-override logic was broken.

I'm not entirely sure this is the best solution to the issue. In the past, you would be unable to send a message to a user in any one-on-one thread, not just `PERSONAL` threads. You can create such a thread (one-on-one, but not `PERSONAL`) by messaging two users, and having the other one leave.

This change only solves the issue for `PERSONAL` threads. To solve it for all one-on-one threads, I could revise the logic to filter non-member out of the `threadInfo.members` checks performed by `threadOrParentThreadIsGroupChat` and `threadOrParentThreadHasAdminRole`. The downside of that approach is that it would be less readable, and in an ideal world the hack won't be necessary for too long. I could be easily convinced to switch to it, though.

Test Plan: Confirm blocking a user on your `PERSONAL` thread with them prevents messaging them (and them messaging you)

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: Adrian, atul, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D1904