[native] Add option to specify drafts height
Summary: Tooltip buttons are going to measure height of drafts and then the hook is going to use these values. This diff allows providing the value and use them to determine the target position.

Test Plan: Modify current and target draft heights and check if the message is animated to the correct position.

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: Adrian, atul, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D1910