[native] Use 150ms transition in all the overlay transitions except message tooltips
Summary: The changed duration was introduced to improve a navigation to a sidebar. Recently, we made sure that when we navigate to a tooltip, we set a navigation key, so now we can check if the key matches and use extended duration only if that's the case.

Test Plan: Open message tooltip and navigate to a sidebar - a longer duration should be used. Open an image modal and check if a shorter duration is used.

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: Adrian, atul, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D1920