[native] Add c++ tests
Summary:
I set up c++ tests on iOS. I think, since the c++ code's mutual, this should be enough. The process is smooth but it requires Xcode(maybe there's a way to launch the tests from the console, I'm using GUI and I didn't check that). For now I added only tests for the crypto module. There are two tests:
- just sending messages between randomly picked "users"(modules) - this simulates a basic communication,
- the case with the [session issue](https://www.notion.so/commapp/Fix-the-logic-of-olm-implementation-sessions-issue-99920460fd124fe58141738165820d5a) when two users send initial messages to each other at the same time and they both initialize outbound sessions.

Of course, we can add some more tests in the future - sky is the limit. But it's a start.

I didn't split this into multiple diffs as a lot of changes here are just configuration files and can be skipped during revision - the rest isn't that much code I believe.

I'm not sure how important is the code quality in the test files as we've not had any test files for c++ so far, so if that matters a lot I can work on that here as I realize the code's a bit dirty.

Test Plan:
Launch the project in the Xcode, go to the `Test navigator`, right-click on the `CommTests` and hit `Run "CommTests"`
Here's the `TestNavigator`:
{F4354}

Reviewers: palys-swm, ashoat

Reviewed By: palys-swm, ashoat

Subscribers: ashoat, palys-swm, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1899