[native] Hide delivery icon while animating to a sidebar
Summary: We would like to hide the icon, but only when we're navigating to a sidebar. We can't simply use overlay position, because the opening tooltip and dismissing it would also hide the icon.

Test Plan: Open a tooltip and navigate to a sidebar - the icon should not be visible. Open a tooltip and close it - the icon should be visible for the whole duration of the animation.

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: Adrian, atul, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D1924