[native] Check if the user appended a text before removing sent message
Summary:
This diff fixes a bug on Android where after sending one letter message and typing it again, that letter was removed from the draft. The issue was caused by the logic which was fixing another bug: when a user types a message, then sends it and starts typing again really fast, the draft contained both: already sent message and the text which the user appended.

The solution checks if the user appended something to the sent message before deleting the already sent part. By doing that we ensure that we're not going to remove the whole new text.

Test Plan:
There were two scenarios to test:
1. Send a letter, type this letter again - the letter should not be deleted.
2. Send some text, quickly start typing again - only the text typed after message was sent should be displayed as a draft.
Tested on Android simulator in release mode.

I also checked if the first issue appeared before the fix and if the second issue appeared when the logic in `onChangeText` was removed.

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: Adrian, atul, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D1936