[lib] Change `RemoveMessagesForThread` operation to be `RemoveMessagesForThreads`
Summary:
In D1927, we're creating a separate `RemoveMessagesForThread` operation for every thread.

Instead, if we have a single op with a set of `threadIDs`, we can delete messages for each of the `threadIDs` in a single SQL query.

In `sqlite_orm` we should be able to do something like (haven't actually tested this code yet)

```
auto remove_messages_for_threads = storage.remove_all<Message>(where(in(&Message::thread, {123, 456, 789})));
```

Test Plan: haven't yet tested

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: palys-swm, Adrian, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D1932