[native] Add option to store chat input bar height in a context
Summary: I decided to use ref here to avoid rendering all the components which use the context every time `ChatInputBar` updates the height. There might be a couple of input bars rendered at the same time - for each thread in navigation, so we need to use a map here.

Test Plan: Tested with next diffs.

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: Adrian, atul, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D1947