[native] Use context value to determine message target position
Summary: Use the current value saved in the context instead of taking this value from tooltips. Only the `targetInputBarHeight` will be provided by the clients.

Test Plan: Added console logs in `ChatInputBar` and in this hook and verified that changing the draft height and opening a tooltip resulted in the same value present in both places.

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: Adrian, atul, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D1950