[native] Allow nullable sidebar input bar height
Summary: The a navigation to a sidebar is started before the bar is measured, we would like to render the animation as correctly as we can. By using the same height of the draft of parent and sidebar, the animation will be correct most of the times. This is a corner case and it shouldn't happen in real life, but still, we need to handle it in the code.

Test Plan: Tested with next diffs

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: Adrian, atul, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D1955