[native] Add option to store simplified animation flag in the context
Summary: To detect if we need to use a simplified approach we need to check if the measurements are finished - and this can only be done in tooltip buttons. The result should be used in `useContentAndHeaderOpacity` and `useDeliveryIconOpacity` so the best way to pass that info is by using the context.

Test Plan: Tested with next diffs

Reviewers: ashoat, atul

Reviewed By: ashoat, atul

Subscribers: Adrian, atul, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D1964