[native] Implement the simplified animation
Summary: In this simplified version we hide a tooltip message and show the message in the message list

Test Plan:
Open a tooltip and navigate to a sidebar where current user is not a member and check if the animation is correct.
Do the same with opened keyboard.
Check if more complicated animation still works e.g. when navigating to a pending sidebar.

Reviewers: ashoat, atul

Reviewed By: ashoat, atul

Subscribers: Adrian, atul, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D1966