[native] fix keyboardAvoidingView behavior on Android
Summary:
I changed keyboardAvoidingView behavior on Android, when the keyboard size changes. For example, while selecting emoticon or gif. The behavior on Android and iOS is now very similar, and the view on Android now always adjusts to keybaord height.
I also adjusted ratchetAlongWithKeybaord behavior on Android, so its config is now the same as in iOS and supports keyboard resizing.

Test Plan: I tested chat view behavior with GBoard on Android (It is Android's default behavior) and on default iOS keyboard, trying to insert and search emoticons and gifs, to make keyboard resize. ratchetAlongWithKeybaord was tested on login and register view. All test were made on virtual devices.

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: ashoat, palys-swm, Adrian, atul, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D1971