[CommCoreModule] Change `Message.id` type from `int` to `std::string`
Summary:
In the messageStore, the IDs for local messages are prefixed with `local`. However, in the SQLite `messages` table, we've given the id column an `INTEGER` "affinity."

Because SQLite has a dynamic type system, we could still insert rows with message IDs that are strings (eg "local100") without issue. However:

A. From what I've read, it's a good practice to have the column affinity match up with the expected contents of the data as much as possible.

B. `sqlite_orm` is strict about data types, which means that to insert a row with a string `id` in the messages table we would need to modify `Message.h`

Migrations 9, 10, and 11 mirror (what were) migrations 6, 7, and 8.

9. Drops existing `messages` table (if it exists)
10. Recreates the `messages` table (if it doesn't exist)
11. Creates a multi-column index on (thread, time)

I "removed" migrations 6, 7, and 8 because they were redundant:

- If an existing client was on version 8 (current), the `messages` table would be dropped and recreated as expected. If we were actively persisting data in the `messages` table, this would be a problem... but we aren't yet.

- If an existing client was on version 7, they would "skip" migration 8, which means they wouldn't have the multi-column index on the previous `messages` table. This doesn't matter because the old `messages` table will be dropped in migration 9 anyway.

- If an existing client was on version 6, they would "skip" migrations 7 and 8, which means they wouldn't have a `messages` table at all. This is fine because migration 9 only drops the `messages` table `IF EXISTS`, and then migrations 10 and 11 can proceed as expected.

Test Plan: Introduced new migrations to client on db version 8: https://blob.sh/atul/int_to_stdstring_migration.png

Reviewers: palys-swm, ashoat

Reviewed By: palys-swm, ashoat

Subscribers: ashoat, palys-swm, Adrian, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D1983