Add thread ancestry properties to ThreadInfo and RawThreadInfo
Summary:
This diff adds `containingThreadID` and `community` properties to `ThreadInfo` and `RawThreadInfo.` (These properties are already on `ServerThreadInfo`.)

They are necessary in order to address [this Notion issue](https://www.notion.so/commapp/ThreadSettings-shows-option-to-add-somebody-who-isn-t-in-community-7ca1e37378ca488ca31925ad52bd6607). The information to answer this question is not fully present on the client currently (see "Is is technically impossible" below for the reason).

The effect of this change will be to cause the state check mechanism to invalidate the contents of every user's `ThreadStore`, which will be costly. In this diff we avoid doing this for outdated clients, since they don't need the new properties.

It is technically impossible to update the `ThreadStore` as a function of the old `ThreadStore` (entirely locally). The reason is that it is possible to be a member of a community (root thread) but not a member of the parent thread, so in some cases there is incomplete information on the client if the parent `RawThreadInfo` is missing from the `ThreadStore`. However, this is an unusual case. In the following diff I'll set up a `redux-persist` migration to transform most of the `ThreadStore`, so the state check mechanism has the do a minimal amount  of work.

Test Plan: Tested in combination with the following diff

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: Adrian, atul, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D1996