[CommCoreModule] Wrap core module threads with unique_ptr
Summary: We need to have an option to start the core module in a single thread mode, where db and crypto operations are run on the same thread. It means that the threads should be initialized later, in which case we should have an option to construct the core module with nullptr threads - and that can be achieved by using the pointers.

Test Plan: Save a draft and check if it is persisted

Reviewers: ashoat, karol-bisztyga

Reviewed By: ashoat

Subscribers: Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D2000