[3/5] [CommCoreModule] Construct `vector` of `Media` from `ClientDBMessageInfo`
Summary:
`ClientDBMessageInfo` is the payload that gets sent from JS to the C++ side.

The `media_infos` property of the payload has a list of `ClientDBMediaInfo` objects. Each of these objects corresponds to a row in the SQLite `media` table.

We will pass this `std::vector<Media>` to the `ReplaceMessageOperation` constructor so the data can be persisted when `execute()` is called.

Test Plan:
1. Passed a few media messages (from my production redux store) to `processMessageStoreOperations`.
2. Set breakpoints to observe the `Media` objects and `vector<Media>` objects were created as expected.

Testing included messages with no/one/some image(s).

Reviewers: ashoat, palys-swm

Reviewed By: ashoat, palys-swm

Subscribers: palys-swm, Adrian, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D2056