[lib] Introduce 'assertThreadStoreThreadsAreEqual' function
Summary:
I added function that compares thread stores to check if the new "ops" pipeline works in the same way as an old one.
The eslint annotation wil be removed in the next diff, when the function will be used.

Test Plan: The change doesn't change any behavior. App on Android emulator runs properly.

Reviewers: palys-swm, atul, ashoat

Reviewed By: palys-swm, atul, ashoat

Subscribers: ashoat, palys-swm, Adrian, atul, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D2039