[CommCoreModule] Modify CommCoreModule so all the jobs are scheduled
Summary: We're going to extract an internal module where all the jobs are run and we will call this module from the `CommCoreModule`. Now, if we call internal module from the core module, the core module will decide how to run it (on a separate thread), but if it's called from other places, the code will execute on a main thread.

Test Plan: Run the app, save a draft and check if it was saved

Reviewers: ashoat, karol-bisztyga

Reviewed By: ashoat, karol-bisztyga

Subscribers: Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D2158